Create initial functions to calculate basic time
Created the calculateTime function which uses a while loop to count the
minutes between two times. The convertTime function is called which
checks the format of the time entered. If the time is hours only it
returns the time, if the time has minutes or uses 24 hour time format
then the convertFormat function is called which slices the input to
record hours and minutes separately. The hours are then converted to
minutes and both values are added together and returned to be
calculated. This is done so the program can recognise different time
formats entered by the user.