Add error handling and overtime logic
Added error handling
- Entering same start and finish time would equal 0. This is now fixed
so that 12 hours are added onto the finish time.
- Fixed toggle for 8 and 10 hours overtime. When toggling between radio
buttons for overtime, it would not deselect the previous buttons
function. This was resolved by assigning false to the opposite flag on
the click event listener

Overtime logic was added and the program now correctly calculates
overtime for 8. 10 and 12 hours respectively.